requirements.txt
# Aetheria 專案 Python 依賴套件

# Google Gemini API
google-generativeai>=0.5.0

# 環境變數管理
python-dotenv>=1.0.0

# Flask Web 框架
flask>=3.0.0
flask-cors>=4.0.0

# HTTP 請求
requests>=2.31.0

# 八字命理计算（寿星天文历）
sxtwl>=2.0.7

# 簡繁轉換（確保輸出為台灣繁體）
opencc-python-reimplemented>=0.1.7

# 測試框架
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-flask>=1.2.0

# 可選：資料處理
pandas>=2.0.0

# 可選：快取（如果需要 Redis）
redis>=5.0.0

# 可選：Web 框架（未來使用）
fastapi>=0.110.0
uvicorn>=0.27.0
